// apps/admin/src/api/generated/model/feedbackConversationResultsDtoOutput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Join The Six API
 * Operations API for Join The Six
 * OpenAPI spec version: 1.0.0
 */
import type { FeedbackConversationResultsDtoOutputAnswersItem } from "./feedbackConversationResultsDtoOutputAnswersItem";
import type { FeedbackConversationResultsDtoOutputNotesItem } from "./feedbackConversationResultsDtoOutputNotesItem";

export type FeedbackConversationResultsDtoOutput = {
  answers: FeedbackConversationResultsDtoOutputAnswersItem[];
  notes: FeedbackConversationResultsDtoOutputNotesItem[];
};